feat: update rating from reviews system on inventary service

+++ apps/merchant-api/InventoryService/src/InventoryService.Domain/Concretes/Product.cs
@@ -10,7 +10,7 @@ public class Product : BaseEntity
     public double BasePrice { get; set; }
     public string Brand { get; set; } = string.Empty;
     public int? LowStockThreshold { get; set; }
-    public int Rating { get; set; } = 0; 
+    public float Rating { get; set; } = 0; 
     public ICollection<Image> Images { get; set; } = new List<Image>();
     public ICollection<ProductVariant> ProductVariants { get; set; } = new List<ProductVariant>();
     public ICollection<Category> Categories { get; set; } = new List<Category>();
