+++ apps/merchant-api/MerchantCommon/Commons/Messages/RatingMessage.cs
@@ -3,5 +3,5 @@
 public class RatingMessage
 {
     public required Guid ProductId { get; set; }
-    public required int Rating { get; set; }
+    public required float Rating { get; set; }
 }
